+++ src/syntactical/syntactical.py
@@ -29,10 +29,11 @@ custom_grammar = r"""
     ?line_content: statement (SEMICOLON statement)* [SEMICOLON]
     ?statement: import_stmt | with_stmt | class_def | func_def | return_stmt 
             | assignment | if_stmt | while_stmt | for_stmt | try_stmt 
-            | break_stmt | continue_stmt
+            | break_stmt | continue_stmt | pass_stmt
             | expression | from_stmt
     break_stmt: "break"
     continue_stmt: "continue"
+    pass_stmt: "pass"
     block: "{" line_content+ "}"
     import_stmt: "import" IDENTIFIER ["as" IDENTIFIER]
     from_stmt: "from" IDENTIFIER "import" IDENTIFIER
@@ -145,9 +146,10 @@ class ToPython(Transformer):
     def iterable_for(self, var, iterable, body):
         return f"for {var} in {iterable}:\n{body}"
 
-    # Break and continue statements:
+    # Break, continue, and pass statements:
     def break_stmt(self): return "break"
     def continue_stmt(self): return "continue"
+    def pass_stmt(self): return "pass"
 
     # Here's the if statement, it's kind of complicated:
     def if_stmt(self, *parts):